[SYCL] Return true in has_kernel for sub-devices
SYCL 2020 specifies that kernel_bundle::has_kernel(kernel_id, device)
should return true if the kernel is valid for a given device. As such
sub-devices where any of its parents are in the kernel bundle should be
considered valid. This commit changes the implementation of has_kernel
to recursively check the parents of a passed sub-device if it is not
itself in the kernel_bundle's list of devices.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>