[Driver] -save-temps issue with files in CWD (#7955)
When using -save-temps against fails in the $CWD, the driver was
avoiding creating the intermediate files in the $CWD and creating them
in /tmp. A previous modification to improve our behaviors with potential
clashing intermediate files when working with llvm-foreach compilations
inadvertently modified the check for the output name matching the base
input name. Put this check back in.