[SYCL] Add property_list parameter to finalize
Introduces an optional `property_list` parameter when creating
an executable graph with finalize.

No properties are defined that could be passed here, but the
kernel fusion API `fusion_wrapper::complete fusion()` takes a property list,
which could be relevant in future. Additionally, CUDA has
a `cudaGraphInstantiateWithParams()` entry-point that this property list would provide
equivalent coverage for.

Using a `property_list` in this way is analagous to how
[kernel_bundle](https://registry.khronos.org/SYCL/specs/sycl-2020/html/sycl-2020.html#sec:interfaces.bundles.overview.synopsis) `compile()`/`link()`/`build()` take a property list
and return a new object with different state.