[SYCL][ESIMD] Don't error on functions with local names
In the ESIMD Verifier, we have a known good list of functions
and want to check that any ESIMD code only calls one of those functions.
We do that by demangling the names of all ESIMD functions  and comparing them
against the known good functions. In the case where there's a function
that is from a local name such as a lambda, functor, or local class,
the demangled name may include the return type of the parent function
if it's a template function. If this happens, we end up using the return type
in the function name check which makes us check the wrong thing.

We don't care about local names, so just skip them.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>