[ESIMD] Fix usage of uninitialized field ESIMDVerifier::ForceStatelessMem
The field was properly initialized in 'ESIMDVerifierPass` struct, but
not in `ESIMDVerifier`.
The first one gets the proper/guarantted value from
LangOpts.SYCLESIMDForceStatelessMem, while the second(i.e. ESIMDVerifier)
could not do the same as it is initialized from
llvm::createESIMDVerifierPass()

Fixing this issue showed that those SYCL methods that are allowed
with -fsycl-esimd-force-stateless-mem, can potentially and legally be used
in some other modes, for example, when used in LLVM IR consumed by 'opt' component.
Thus the meaning and name of the corresponding fields in ESIMDVerifier and
ESIMDVerifierPass structs were renamed from ForceStatelessMem to
MayNeedForceStatelessMemModeAPI.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>