[SYCL] Add device_global initialization and PI APIs (#7797)
This commit adds initialization of device_global without
device_image_scope. This is done as follows:
1. Allocate USM memory for the device_global.
2. Fill the USM memory with zeros. Store this event as a dependency for
related kernels.
3. When enqueuing a kernel using one or more device_globals, get the
initialization events for each of them (unless they have finished) and
add them as dependencies for the kernel launch.

Additionally, to write the USM pointer to the device_global, PI has two
new functions: piextEnqueueDeviceVariableWrite and
piextEnqueueDeviceVariableRead. These allow for writing to and reading
from device_globals in a given program. These are implemented for
OpenCL, L0, and CUDA. For CUDA sycl-post-link will generate a mapping
from the device_global's unique ID and the name in the binary.

Design doc:
[sycl/doc/design/DeviceGlobal.md](https://github.com/intel/llvm/blob/sycl/sycl/doc/design/DeviceGlobal.md)

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>