Translate aliases to aliasee value in more contexts (#1809)
Currently llvm-spirv crashes with an assert when translating uses of
aliases which are not the callee of a function call instruction. This
change remedies this by handling these uses of aliases the same way
function call callees are handled: by using the value of their aliasee.

Uses SPV_INTEL_function_pointers when the alias is used as a function pointer

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/0378393