Use correct template args for host_accessor constructors
Prior to this change using the host_accessor zero dimension constructor
would result in a series of errors because some constructors were using
`DataT, Dimensions` (the class template args) for `buffer`. During
template exploration, the compiler would try and instantiate buffer
with `Dimensions = 0`, which is not possible, and that would trigger
the series of errors mentioned before. Using `T, Dims` (the function
template args) the rest of constructors are not enabled because the
dimensions of the host_accessor and the dimensions of the buffer don't
match.