[SYCL][Reduction] Fix last group detection's atomic memory_order
It needs to be at least `memory_order::acq_rel` because we need updates
to the partial sums to be visible to the last finished thread.

This was caught on an existing test from the intel/llvm-test-suite that
we've run internally on a new configuration, hence no complementary test
suite change.