[SYCL] Add cmake support for the SYCL End-to-End tests
We plan to move the intel/llvm-test-suite back in-tree. In order to do
this we need to update the infrastructure first (e.g., CI). Enable that
by adding require CMake support and a "fake" test.

We plan to run the tests in a "standalone" mode, similarly how current
intel/llvm-test-suite is used. As such, CMake target is decoupled from
general check-* processing and is not include into check-all
automatically.