[SYCL] Enable DPC++ build with llvm-mingw toolchain on Windows (#7859)
This patchset fixes some issues found when trying to compile DPC++
compiler on Windows with Clang in MinGW+ucrt environment. It basically
swaps 'if(MSVC)' with 'if(WIN32)' in a few places to make it build the
correct result.

With this patchset I have managed to build a somewhat working compiler
that successfully builds oneAPI code examples.