[SYCL] Throw exception when selected device is not in context (#8052)
In accordance with the SYCL 2020 specification a device selector should
always pick from all available devices. Likewise, the queue constructor
should throw an exception if the selected device is not in the given
context. This commit fixes the device selection behavior when a context
is specified, throwing the corresponding exception. Note that this only
affects SYCL 2020 style device selectors.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>