[SYCL] Move backend macro definition to plugin headers (#8008)
This patch introduces feature headers provided by the plugins to define
the backend macros. The main feature test header can then use
`__has_include` to conditionally include the plugin headers.

This is better than the current system because it allows to install a
plugin build in an existing DPC++ installation by just adding the plugin
feature header and the plugin shared library without having to touch the
existing feature test header.

It may also be good as a follow up to move backend interop and traits
headers into the plugins and use a similar system to install and
distribute them.

As a side note, what we're doing with the backend traits and backend
macros is technically incorrect with the way the SYCL spec is currently
written, as it would mandate that if a backend is listed in the backend
enum class, it must have defined backend traits and a backend macros,
but in DPC++ the values are still in the enum when the plugins are not
built. I've suggested a change to the SYCL spec to relax this:
https://github.com/KhronosGroup/SYCL-Docs/pull/343