[SYCL] Refactor program build and cache
In order to help device_global find associated programs in the cache,
the way the program manager and the cache interacts needs a refactoring.
This commit makes this refactoring.

Additionally this adds the notion of a "common key" (CommonProgramKeyT)
representing the common parts of the full cache key shared between
builds of the same program on the same device, excluding specializations
such as specialization constants and compilation options.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>