Fix reverse translation of ptr.annotation for multiple UserSemantic decorations (#1723)
It was found that the decorated variables (results of ptr.annotation intrinsic) are not used in further instructions unlike the input source. The issue is present for the reverse translation of multiple UserSemantic decorations.

This change fixes the reverse translation of multiple ptr.annotation intrinsics on the same class field. With this, the result of intrinsic is not lost and actually reused in further intrinsics and instructions.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/308daff