Throw correct exception when passing unbound accessor to command
According to 4.7.6.9 of SYCL2020 spec, if a placeholder accessor
is passed to a command without being bound to a command group, an
exception should be thrown. Prior to this change, the exception
was only thrown if an unbound placeholder accessor is used, not
only passed. Also, prior to this change, if we passed one unbound
accessor and one (or more) bound accessors to a command, the
exception thrown was incorrect.