Add new intrinsics and attributes to control accuracy of FP calls
This patch adds a new set of fpbuiltin intrinsics to represent operations
that are equivalent to common math library functions and basic operations,
and adds a new call site attribute ("fp-max-error") to specify the required
accuracy of these calls.

The purpose of these new IR constructs is to support alternate math library
implementations and provide a general mechanism for selecting among multiple
implementations based on specific requirements.