[SYCL][CUDA] Fix context setup for device infos
Extend the `ScopedContext` to work with just a device, in that case it
will simply use the primary context.

This is helpful for entry points that only have a `pi_device` and no
`pi_context` but that still need some cuda calls that require an active
context, such as for the device infos.

This addresses a bug where getting the amount of free memory before
creating any queues or context, would simply crash.

This was partially solved in a previous PR, however the previous PR was
releasing the primary context, but leaving it active on the current
thread, so getting the device info twice in a row would end up crashing
again since it would just use the active but released primary context.