[SYCL][Reduction] Fix reduction kernel conflict
Certain reductions have to check for the atomic64 aspect on a device at
runtime before selecting between different kernels. This means both
kernels must be generated, but currently this causes two different
reduction kernels with the same name to be generated.
This commit adds a new reduction kernel name wrapper (AspectDepKrn) used
by the reduction selected when the atomic64 aspect is supported to avoid
the conflict.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>