[SYCL] Change signature of host_accessor::get_pointer for SYCL 2020 (#8109)
SYCL 2020 specifies that the `get_pointer` member should now return a
raw pointer for based on `value_type` and should be `noexcept`. This
commit changes the signature of `get_pointer` for host accessors.

Note that `AccessorBaseHost::getPtr` cannot be made `noexcept` currently
as it would be an ABI break. This does not limit the functionality of
these changes, but should be changed during next ABI break.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>