[SYCL][NFC] Do not use non-standard aliases in images impl
SYCL 2020 spec do not contain definition of aliases like `cl_float4` and
therefore we can't use them in our implementation.

The lack of such aliases may be a mistake in the spec, but SYCL 2020
images do not reference any `cl_*` types anyway, so the assumption is
that at some point we should be able to switch to aliases like `float4`
(once we drop SYCL 1.2.1 support).