Fix extensions/INTEL/SPV_INTEL_function_pointers/alias.ll for intel/llvm
The intel/llvm version of llvm-spirv generates entry points differently
than upstream; based on my understanding, this is to work around an IGC
bug. This causes a SPIR-V mismatch for the bar entry point in this
testcase. bar doesn't need to be an entry point for the purposes of this
test, so this change resolves this test failure in intel/llvm by making
bar a normal function.