[SYCL][NFC] Do not use non-standard aliases in image utils (#8139)
SYCL 2020 spec do not contain definition of aliases like `cl_float4` and
therefore we can't use them in our implementation.

The lack of such aliases may be a mistake in the spec, but SYCL 2020
images do not reference any `cl_*` types anyway and modified code is
mostly for images support on host. It should be fine to switch to new
SYCL 2020 aliases and fixed width integer types because they provide the
same size guarantees as their `cl_*` counterparts.