[SYCL] Separate CompileTimePropertiesPass from sycl-post-link tool (#7527)
Extracted the `CompileTimeProperties` pass from the `sycl-post-link`
tool up into the `llvm` directory with the rest of the passes and added
it as a step in the code generation pipeline. Moving this into a more
generalised location will allow for new tools to be developed that
require its logic.

Due to a dependency on `DeviceGlobals`, that also had to be moved and
includes adjusted in `sycl-post-link` tool. This pass can be invoked via
opt as well for testing.

Several `sycl-post-link` tests had to be transferred to
`llvm/test/SYCLLowerIR` with some being modified as well to account for
the change. Some tests had to be split across the two directories so as
to maintain coverage. Particular tests that focused on the result of
running the pass over some iput bytecode and testing the result have
been turned into their own tests under the
`llvm/test/SYCLLowerIR/CompileTimePropertiesPass` test directory.