[SYCL] Add sycl-single-task implict property on single_task
This commit adds the "sycl-single-task" LLVM IR attribute to all
SYCL kernels originating from single_task calls. Leveraging the existing
mechanism for creating attributes for the SYCL kernel compile-time
properties extension, applying the add_ir_attributes_function to SYCL
kernel wrappers, this new attribute acts like a kernel compile-time
property.

To avoid this implicit attribute from causing property/attribute
conflict warnings, a special rule is made in the frontend to skip the
conflict check if it is the only value in the add_ir_attributes_function
attribute.

Additionally, this commit adds special handling of "sycl-single-task"
when targeting "spir64_fpga", in which case it will cause a
"max_global_work_dim" metadata node to be created with value "0" on
functions where it is present.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>