[SYCL] Implement host pipe unique name generation and mapping calls (#8009)
Implementation of host pipes outlined in the design document in this PR:

https://github.com/intel/llvm/pull/5850

1. Generation a unique pipe id for GVs marked with the new
"sycl-host-pipe" attribute. Id generation utilizes the same method as
used for name generation for device global.
2. Added a host pipe map to map the addresses of marked GVs with the
unique id. This host pipe map is generated by a constructor and method
calls added to the header and footer.
3. Modified the sycl-post-link tool to generate compile time properties
metadata for these GVs. This metadata contains the unique id generated
for the GV to be consumed by the device backend compiler.

PR for accompanying runtime changes:
https://github.com/intel/llvm/pull/7468

---------

Co-authored-by: Alexey Sachkov <alexey.sachkov@intel.com>