[SYCL] Usm allocator refactor (#7785)
This commit makes USMAllocator unaware of the memory type on which it is
operating. The user of USMAllocator is responsible
for setting appropriate options for each instance of USMAllocator.

This is the first step to making USMAllocator generic. In the next step,
I'd like to start integration with Unified Memory Allocation
(oneMemory). USMAllocator would be used as a heap manager by UMA.