[VectorUtils] Enhance VFABI demangling API
This is a cherry-pick of https://reviews.llvm.org/D141650

This patch adds more flexibility to the original VFABI demangling API.

1. Allow demangling an empty parameter list, which is explicitly allowed by
X86 VFABI doc (section 2.6) and probably implicitly allowed by AArch64 VFABI
doc (section 4.1) as well.
2. Allow passing the module as a nullptr, so that we don't check the existence
of the demangled vector function name. This is useful when we want to create
the widened function according to the demangling result.

Signed-off-by: Yilong Guo <yilong.guo@intel.com