[SYCL] Enable stack printing on crashes in post commit (#7934)
The initial attempt to enable LLVM stack printing on crashes(SEGFAULT)
when running SYCL tests.
The motivation is to simplify analysis of sporadic fails in CI.

The patch:
1. If `DSYCL_ENABLE_STACK_PRINTING` `cmake` var is defined:
a. Links the `libsycl.so` with `LLVMSupport` which provides needed
functionality
b. Calls `llvm::sys::PrintStackTraceOnErrorSignal`, which registers
signal handlers,
        in a couple of places in SYCL RT
    c. Adds `llvm-sybmolizer` to the dependencies of `sycl-toolchain`
2. Modifies post-commit workflow to:
    a. build linux job (`linux_default` only) in RelWithDebInfo mode
    b. pass the `cmake` var to the `configure.py`
3. Sets `LLVM_SYMBOLIZER_PATH` in the `lit.cfg.py` for SYCL to the
`llvm-sybmolizer` binary from the build dir