[SYCL] Report profiling as not supported if submit timestamps are not
With the recent fix to command_submit profiling information query it now
requires that the plugin supports piGetDeviceAndHostTimer API. Because
of that, we should not report queue profiling as a supported aspect
unless both the underlying pi_queue supports profiling and
piGetDeviceAndHostTimer is available.

In addition, add a feature_not_supported exception when a queue is
constructed with enable_profiling property and a device that does not
support profiling.