[SYCL] Add sycl kernel arg annotation support (#6904)
Added new sycl kernel arg annotation properties to sycl-post-link and
spirv translator. This PR is pending on spec to be settled.

Dependency:
A SPIRV header change is required to be pulled down. The Pull request is
here
https://github.com/KhronosGroup/SPIRV-Headers/pull/291

A SPIRV translator change is required to be pulled down. The Pull
request is here
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/pull/1632

The two PR above are merged. But it seems the pull down from Khronos has
some problems, thus manually merge the SPIRV-Headers and
SPIRV-LLVM-Translator here
https://github.com/intel/llvm/pull/8120

---------

Co-authored-by: Alexey Sachkov <alexey.sachkov@intel.com>