[SYCL] Fix variadic marray ctor (#8271)
Currently the variadic marray constructor only accepts values that are
directly convertible to the element type of the marray. However,
according to the SYCL 2020 specification the arguments of the
constructor can be any combination of such values and marrays with
suitable types, where the aggregate size of the arguments must match the
number of arguments of the specified marray. This patch allows the use
of marray arguments in this constructor.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>