[SYCL][CUDA] Move base event into the device
Events are tied to a specific context, calling `cuEventElapsedTime`
with two events created in separate contexts returns a
`CUDA_ERROR_INVALID_HANDLE` error.

So it really doesn't make sense to keep that on the platform, and now
that we're using in the primary context and tied it to the device, it
makes sense to also move the event base into the device class.

This is technically not a problem right now because we only have one
device per platform, but it's a bit cleaner will help towards the
multi-device context work.