[SYCL][Fusion] Enable fusion of kernels with different ND-ranges (#8209)
All kernels with the same (or unspecified) local size and offset can be
fused. In order to make this work, some builtins getting index space
information must be remapped and the resulting ND-range of the fused
kernel, calculated.

The ND-range of the fused kernel will have:
1. The same number of dimensions as the input ND-range with the higher
number of dimensions;
2. The same local size as the shared local size (or unspecified)
3. The same offset as the shared offset
4. The global size will be the **greatest** input global size as per the
following ordering:
  i. Number of work items (enforces correctness);
  ii. Number of occurrences (less remappings needed);
  iii. Lexical order of the dimensions (introduces determinism).

Builtins obtaining the local/global size/id, work-group id, number of
work-groups or offset are remapped introducing as per an alwaysinline
function that can be reused along the fusion pass. More information can
be found in the Builtins.cpp file, where the remapping logic is
implemented.

---------

Signed-off-by: Victor Perez <victor.perez@codeplay.com>