Updated aspects list to be SYCL 2020 compliant
- Removed usm_restricted_shared_allocations aspect
It is not a part of the SYCL 2020 specification so has been removed.

- Deprecate host aspect in favour of cpu as "host" device concept removed from SYCL 2020
Updated tests to add diagnostics for deprecation warning where needed.

Once the host aspect is to be removed, these tests can go back to
not requiring diagnostics.

- Added placeholder aspects for emulated and host_debuggable
Currently both return false while while further implementation details are obtained.

- Removed duplication host_debuggable aspect def

- Added emulated aspect to device_has/have