[libclc] Prefer StringRef fixing string handling bugs (#8296)
In addition to slightly reducing the cognitive overhead of all the
different string types, this fixes two real-live bugs:
1. `std::atoi` is impossible to use correctly. We'd incorrectly assume
invalid address spaces mangling as AS0 because `atoi` return zero on
error. This has been replaced with `StringRef::getAsInteger`
2. A crash handling nested names due to incorrect use of
`string::find_first_not_of("__spv::")` which searches for the first
character that is not one one of the list of characters given by the
first argument. The intention was to split on that substring, which
`StringRef` has easy interfaces for achieving. Names in nested
namespaces can now be parsed without a crash