[SYCL] Fix handling of host-side memory in 2D memops
The current implementation of 2D USM memory operations will launch a
kernel as a fallback mechanism. However, these interfaces are also
intended to accept pointers USM and non-USM host memory. To allow this,
the implementation now picks different fallback mechanisms depending
on the type of pointer passed.

For fill/memset this means:
 1. Always use backend-defined interface if available.
 2. Use auxiliary fill kernel if pointer is to device memory.
 3. Use a host-task with a fill per row if the pointer is to host
memory.

For memcpy/copy this means:
 1. Always use backend-defined interface if available.
 2. Use auxiliary copy kernel if both pointers are to device memory.
 3. Use a host-task with a copy per row if both pointer are to host
memory.
 4. Enqueue a USM memcpy per row if one pointer is to in host memory and
the other pointer is in device memory.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>