[SYCL] Add support for the in-tree end-to-end tests
We plan to move the intel/llvm-test-suite back in-tree. In order to do that, we
need to update the infrastructure first, both in this repository and internally.
Start by adding a single trivial test and enabling cmake/Linux CI support.

We plan to run the tests in a "standalone" mode, similarly how current
intel/llvm-test-suite is used. As such, CMake target is decoupled from general
check-* processing and is not include into check-all automatically.