[SYCL][Reduction] Limit reduction work non-uniform case
In certain cases non-uniform reductions will not have enough work to
fill the partial reduction in the last work-group. Previously we would
pad with the identity to work around this, but to avoid needing to do
this we can now instead reduce the size of the partial reduction to the
actual amount of work.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>