[SYCL] Use `dim_loop` to unroll loops in `reduce_over_group` in cuda backend. (#7948)
A performance regression was reported when using `reduce_over_group`
with sycl::vec.
This was due to a loop over calls to the scalar `reduce_over_group` for
each of the `sycl::vec` components that was not unrolled and led to
register spills even at -O3.
It was initially possible to fix the performance by calling `#pragma
unroll` and declare `reduce_over_group` with
`__attribute__((always_inline))`. However the `SYCL_UNROLL` macro that
calls `#pragma unroll` has been removed in favour of `dim_loop`
(https://github.com/intel/llvm/pull/6939).
I have used dim_loop to fix the loop unrolling. However, in the cuda
backend, just using `dim_loop` in this way actually makes the
performance worse. This is because `dim_loop` introduces new non inlined
function calls in the cuda backend that lead to register spills. The
solution to this coincides with the solution of several user reports
that the cuda backend is not aggressive enough with inlining. In this PR
I have also therefore increased the inlining threshold multiplier value
to 11.

See https://reviews.llvm.org/D142232/new/ for the corresponding upstream
PR (for the inlining threshold change) that includes much more details
on benchmarking dpc++ cuda with this change. In short, for dpc++ cuda,
there is no other downside apart from a very small increase in compile
time in some cases, but there is a massive benefit to increasing the
inlining threshold across a large amount of applications.

Testing using opencl cpu backend reveals that this code change has no
effect on this backend. This change is required for the cuda backend but
should have no performance effect for other backends.

fixes https://github.com/intel/llvm/issues/6583.

---------

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>
Co-authored-by: JackAKirk <chezjakirk@gmail.com>