[SYCL] Fix sycl-post-link action output type for non-SPIRV (#8394)
With HIP, we use the output of clang-offload-bundler directly in target
linking. Usually this is fine, because it becomes an input action for
linking all device inputs together, and the link action has IR output
type. However, for -fno-sycl-rdc we don't link the device code together,
so the clang-offload-bundler action is a direct input to sycl-post-link.
Since sycl-post-link previously copied the output type of the input
action, we would end up with Object output type, which caused '-x
object' to be pased to the HIP compiler, which is wrong. Instead, mark
sycl-post-link as always outputting IR for non-SPIRV.

Closes: https://github.com/intel/llvm/issues/8364

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>