[SYCL][XPTI] Fix using xptifw in external projects (#8436)
Currently, xptifw, when built outside of SYCL, downloads its own copy
of the googletest framework. This leads to target name collisions
when xptifw is included in other CMake projects (e.g., unified runtime),
so add an option to disable tests in xptifw.

This project also hardcodes the location for the XPTI proxy
implementation. This prevents its use outside of SYCL, so allow setting
external xpti dir in xptifw.