[SYCL] Temporarily disable strict fp64 checks (#8375)
With the optional kernel feature checks enabled the runtime has become
stricter when checking for aspect requirements. As a result of this, any
use of double floating points, including those optimized out by the
compiler, are being treated as incompatible with devices that do not
support fp64. This means existing user code that relied on this could
break when trying to run on devices that do not support fp64.

This commit disables the aspect check for fp64 temporarily, returning
the control for this check to the backends while we investigate how
better to address this.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Alexey Bader <alexey.bader@intel.com>