[CMake] Avoid configuring headers into sources
We have an arrangement where CMake recursively globs to copy headers
directly from the sycl/include/ source directory both into the build
directory and for installation.

This change moves unconfigured (i.e., *.hpp.in) sources out of this
directory so that they're not copied into the build area nor installed
as part of the globbing arrangement. Instead, they're configured
directly into the build area and installed from there.

This ensures that builds don't modify the source tree and also avoids
unnecessarily copying or installing the *.in files.