[SYCL][L0] Don't append commands to the closed cmd list (#8447)
Shortly, problem is that currently we can't call executeCommandList for
the closed command list. If we do this then it is marked as "open" again
and new commands are appended there which leads to a hang.

Details:
To insert a queue-wide barrier in the piEnqueueEventsWaitWithBarrier we:
1. get a command list for each level zero queue in the pi_queue. We
insert barrier into each command list and each inserted barrier signals
its own event.

2. We create an event list using createAndRetainPiZeEventList based on
events from all barriers -> At this point some of the open command lists
may get closed and executed.

3. We insert convergence barrier into the first cmd list.

4. We call executeCommandList for each command list and that will lead
to the fail if command list was closed at step 2. Because that command
list will be marked as "open" again.