[SYCL] Fix atomic_ref ctor to fix operator ambiguity (#8295)
The `atomic_ref` constructor taking a reference to its datatype should
be explicit, but currently the implementation of `atomic_ref` for
pointer types is not. This causes ambiguity in the overloaded assignment
operator as the non-explicit constructor makes overload resolution find
the deleted assignment operator. To fix this, this commit makes the
`atomic_ref` constructor explicit as intended.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>