[SYCL][Reduction] Refactor tree-reduction to avoid identity (#8424)
This commit refactors the current implementation used in reductions when
doing a tree-like work-group reduction. The current implementation
relies on both the identity to be present and the reduction buffer to
have an additional element to be used for intermediate results. The new
strategy first makes the reduction into a power-of-two reduction by
computing partial sums for the greatest power-of-two smaller than the
total number of elements, then reduces them using the knowledge that the
remaining work is a power-of-two to do proper balanced tree reduction.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>