[SYCL-MLIR]: Create SYCL AliasAnalysis and use it in polygeist LICM pass (#8430)
The MLIR local alias analysis is very conservative. This PR introduces
an AliasAnalysis focused on SYCL dialect operation. The intent is to
augment the basic MLIR `LocalAliasAnalysis` with the new analysis.

This PR introduces the new `sycl::AliasAnalysis` class, with initial
support for `sycl.accessor.subscript` operations.
The new analysis is used in the polygeist LICM pass to allow that pass
to hoist loop invariant `sycl.accessor.subscript` operations.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>