[SYCL] Default initialize group_local_memory_for_overwrite
In accordance with the extension introducing it, the
group_local_memory_for_overwrite should default initialize its data.
This commit adds this behavior, making it only initialize and
synchronize when needed.

Additionally it changes the local ID calculation in group_local_memory
use the group, as suggested in the TODO.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>