ur: Introduce urKernelSetSpecializationConstants
The issue #269, and the PR intel/llvm#7946, outline a desire to enable
adapters to opt-in to setting specialization constants on a kernel
object rather than a program object. On these adapters it is not
required to perform compilation or linking when setting specialization
constants and as such they can be set immediately prior to
`urEnqueueKernelLaunch` with very low overhead.

This patch:

* Adds the `UR_DEVICE_INFO_KERNEL_SET_SPECIALIZATION_CONSTANTS` device
  query to determine if `urKernelSetSpecializationConstants` is
  supported by the device.
* Adds `ur_specialization_constant_info_t` to represent the data
  required to set an individual specialization constant
* Adds `urKernelSetSpecializationConstants` entry point, which adapters
  can opt-in to supporting, which enables setting multiple
  specialization constants on a kernel object.
* Replaces the singular `urProgramSetSpecializationConstant` with the
  plural `urProgramSetSpecializationConstants` to enable setting
  multiple specialization constants at once on a program object and have
  a consistent API between the program and kernel variations.