[SYCL][ESIMD] Make sure metadata references new function in ESIMDLowerVecArg
In ESIMDLowerVecArg, we change the type of a function and fix uses to use the new one.
However, we forgot to handle metadata, so any metadata referencing the old function
would end up as null after this pass, as we also delete the function. Make sure
to fixup metadata uses too.

Note this pass only runs with opaque pointers off, so we may not need it much longer.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>