[SYCL] Windows Proxy Loader for DLLs (#8242)
DLLs manually loaded by SYCL are not tracked as direct dependencies in
the same way that linked DLLs are. This means these DLL may be unloaded
before SYCLs shutdown() routine is called, which will lead to problems
when that routine tries to call those DLL to release resources. This PR
adds a new proxy DLL that is a linked dependency of SYCL itself. This
proxy DLL loads all the SYCL manually loaded DLLs early, before SYCL
itself is loaded, and conversely, they are not unloaded until the proxy
itself unloads, which is after SYCL unloads. So now the manually loaded
plugin DLLs will be resident when shutdown() is called and piTearDown
can complete safely and successfully.

I had a previous PR for this work (
https://github.com/intel/llvm/pull/7756 ), but it encountered
interference with a difference in how Windows handles threads and their
termination. I'm addressing that problem separately. In this version, I
am reducing the shutdown() procedure on Windows to only release the
plugins and nothing else. This avoids the issue for now.

Tests are at https://github.com/intel/llvm-test-suite/pull/1465

---------

Signed-off-by: Chris Perkins <chris.perkins@intel.com>