[SYCL][sycl-post-link] Make GroupId a string to avoid dead references
In certain cases the EntryPointGroup may outlive the string reference
passed in its specified name, in which case its GroupId member dies
prematurely. This can cause uses of the dead string reference. To avoid
this, this commit makes the GroupId member into a copy of the passed
string instead of a reference.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>