[SYCL-MLIR]: Adjust use of the Pure interface in the SYCL MLIR dialect (#8441)
Several operations in the SYCL dialect currently use the `Pure`
interface on the operation. The MLIR `Pure` interface is defined as:

> // Always speculatable operation that does not touch memory. These
operations
>   // are always legal to hoist or sink.
>   def Pure : TraitList<[AlwaysSpeculatable, NoMemoryEffect]>;

where `NoMemoryEffect` is defined as: 

>  def NoMemoryEffect : MemoryEffects<[]>;

These SYCL operation also define the memory effects on several of their
input arguments as `MemRead`. This is a contradiction because the `Pure`
interface on an operation asserts that the operation does not access
memory. This PR remove the `Pure` interface in those cases.

In addition the PR adds `Pure` to the 'grid' operations in the dialect.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>