[SYCL][CMake] Disable win_proxy_loader build and link on non-Windows
Currently the win_proxy_loader is built and linked to the SYCL runtime
on all platforms, but only carries meaning on Windows. Currently this
causes issues as the corresponding library is not installed on
non-Windows targets. Instead of installing it on non-Windows targets,
this commit makes it only build and link to the SYCL runtime library
when building on Windows.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>