[SYCL] Fix refcounting for kernel_bundle interop with OpenCL (#8465)
OpenCL interop spec basically says that each "conversion" must adjust
reference counter of the underlying object. Many of the interop
interfaces had been fixed in https://github.com/intel/llvm/pull/3236 but
the majority of the kernel_bundle feature was implemented after that and
made the same omission.

I'm not sure if that is the best way to fix it though. Another
alternative would be to move all these retains into the OpenCL PI
plugin.