[uma] Implement memory tracking
Implement memory tracker and expose umaPoolByPtr.

An instance of memory tracker is created and managed by the UMA
library. To interact with the tracker, user can use a
'memory tracking memory provider' that will register all
allocations/deallocations with the tracker.

Also implement umaFree.

This function works similarly to umaPoolFree but does not require
a pool handle. Instead it realies on memory tracking and umaPoolByPtr
function.